requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pdfplumber>=0.11.0
groq>=0.9.0
python-dotenv>=1.0.0
pytest>=9.0.0
